Move code for determining new partition divisions using quantiles to a new file.
